backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
openai==1.12.0
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.31.0
arxiv==2.1.0
semanticscholar==0.7.0
pydantic==2.6.1 